LF-4766 Create separate types file
