+++ packages/api/src/util/ensembleService.types.ts
@@ -1,0 +1,89 @@
+/*
+ *  Copyright 2025 LiteFarm.org
+ *  This file is part of LiteFarm.
+ *
+ *  LiteFarm is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  LiteFarm is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
+ */
+
+interface Point {
+  lat: number;
+  lng: number;
+}
+
+export enum PlantingMethod {
+  BED_METHOD = 'bed_method',
+  CONTAINER_METHOD = 'container_method',
+  BROADCAST_METHOD = 'broadcast_method',
+  ROW_METHOD = 'row_method',
+}
+
+type MethodDetails = {
+  planting_management_plan_id?: string;
+};
+
+interface PlantingManagementPlan {
+  planting_method: PlantingMethod;
+  bed_method: MethodDetails | null;
+  container_method: MethodDetails | null;
+  broadcast_method: MethodDetails | null;
+  row_method: MethodDetails | null;
+}
+
+export interface ManagementPlan {
+  management_plan_id: string;
+  crop_management_plan: {
+    seed_date: string;
+    planting_management_plans: PlantingManagementPlan[];
+  };
+  crop_variety: {
+    crop: {
+      crop_common_name: string;
+      crop_genus: string;
+      crop_specie: string;
+    };
+  };
+}
+
+export interface LocationAndCropGraph {
+  farm_id: string;
+  name: string;
+  location_id: string;
+  figure: {
+    area: {
+      grid_points: Point[];
+    };
+  };
+  management_plans: ManagementPlan[];
+}
+
+export interface EnsembleCropData {
+  crop_common_name: string;
+  crop_genus: string;
+  crop_specie: string;
+  seed_date: string;
+  planting_details?: {
+    planting_method: PlantingMethod;
+    [key: string]: unknown;
+  };
+  management_plan_id?: string; // For dev purposes
+}
+
+export interface EnsembleLocationAndCropData {
+  farm_id: string;
+  name: string;
+  location_id: string;
+  grid_points: Point[];
+  crop_data: EnsembleCropData[];
+}
+
+export interface OrganisationFarmData {
+  [org_uuid: string]: EnsembleLocationAndCropData[];
+}
